Implemented Online Help System Prototype
==============================================================================

-Implemented OnlineHelpSystem.java Activity that hosts various "help topic" fragments that provide
instructions to the user on how to navigate and use the Knight-Ranker Application.

-Uses drawer navigation as a means of selecting help topics.

-Loads individual fragments that elaborate on the specifics of each topic.

==============================================================================

Additional Changes:

-Fixed Leaderboard fragment and other activities/fragment not reflecting appropriate color changes
from color picker.

-Default color for Toolbar = RED and Default color for Background = WHITE.
(yellow was admittedly a nasty color to set both to as the dfeault)

==============================================================================

TODO: Find satisfactory way of displaying a numbered or bulleted list of text via some widget or another.
(A single TextView isn't the best method)

TODO: Implement JavaDoc header comments for each .java file and for each method within each class.

FIXME: None so far.