Implemented additional JavaDoc Headers...
-Added JavaDoc Headers for files that were added to the Online Help System and for files related to the
ConfirmCheckboxActivity.java class.

TODO: Re-name ConfirmCheckboxActivity.java and associated dependencies to be less literally descriptive and more
reflective of its general purpose in the context of the Knight-Ranker Application.
(i.e. "Resolve Matches", "Confirm Matches", "Resolve and Confirm Matches", etc.)

Note: JavaDoc Header comments may not be fully accurate and descriptive of the purpose of the class as I, the committer,
did not actually write these classes and am making an educated guess as to their general function.

Note: To team members, for whatever methods, classes, etc. you wrote, it would be most helpful for code documentation to
document as you go or to go back and elaborate on the purpose of that segment of code.  I can only do superficial analysis
and make an educated guess as to how your code functions. (time crunch ftl)